Fix bug in the parser.
Previously, SPS and PPS NALUs were immediately converted into
`DecoderInstructions` in the parser, while slices (both P and I) were
pushed into `AuSplitter`. This meant a reorder was possible:

1. P frame pushed into `AuSplitter`
2. SPS and PPS returned as decoder instructions
3. I frame pushed into `AuSplitter`, which detects that the I frame
   belongs to a different access unit and return the P frame

In this case, the order was turned from P, SPS, PPS, I into SPS, PPS, P,
I, which is problematic, because the P-frame could be referring to
previous SPS, which are overridden before the P frame can be decoded.

This patch pushes SPS and PPS through the `AuSplitter` to avoid such
cases.