Fix the V4L2 input detecting resolution change every frame
This issue used to happen when using the OBS virtual camera as an input
device. Obs would sometimes add extra 2048 bytes at the end of the
buffer it sends us, which would cause us to detect a resolution change.

Now, resolution change is detected when the difference of the received
buffer length and the expected length is greater than 1% of the received
length.